[Example 2] Moving and Restructuring (#233)
