+++ example_2/description/urdf/diffbot.materials.xacro
