+++ ros2_control_demo_hardware/ros2_control_demo_hardware.xml
@@ -6,13 +6,6 @@
       The ROS2 Control minimal robot protocol for a system with multiple command and state interfaces.
     </description>
   </class>
-  <class name="ros2_control_demo_hardware/DiffBotSystemHardware"
-         type="ros2_control_demo_hardware::DiffBotSystemHardware"
-         base_class_type="hardware_interface::SystemInterface">
-    <description>
-      The ros2_control DiffBot example using a system hardware interface-type. It uses velocity command and position state interface. The example is the starting point to implement a hardware interface for differential-drive mobile robots.
-    </description>
-  </class>
   <class name="ros2_control_demo_hardware/RRBotSystemWithSensorHardware"
          type="ros2_control_demo_hardware::RRBotSystemWithSensorHardware"
          base_class_type="hardware_interface::SystemInterface">
